requirements.base.txt
aiohttp
autoflake
autoimport
beautifulsoup4
bottle
bottle-session
djlint
duckduckgo_search
gunicorn
html2text
jinja2
loguru
lxml
lz4
marko
mypy
ollama
openpyxl
py-d2
pydantic
pyflakes
pytest
python-dotenv
rq
tabulate
tiktoken
